Put common widget functions into utils and scale history graph with 0 as min
resolves #21
