+++ src/widgets/subs.ts
@@ -1,6 +1,7 @@
 import { Subscription } from '../types/subscription';
 import { Config } from '../types/config';
 import { WidgetData } from '../types/widgetData';
+import { getNormalizedValues, startTask, stopTask } from '../app/utils';
 
 const io = require("socket.io/client-dist/socket.io.min"); // use socket.io/client-dist instead of socket.io-client because streamlabs requires older client version
 
@@ -71,53 +72,6 @@ const updateConfig = (config: Config) => {
  */
 const setSubCount = (count: number) => {
   document.querySelector('p').innerText = `${count}`;
-};
-
-/**
- * Run a task at a certain interval.
- * If it is already running, stop the old interval first.
- *
- * @param intervalId the intervalId for this task
- * @param task the task/function to execute
- * @param interval the interval to run the task (in seconds)
- * @returns {number}
- */
-const startTask = (intervalId: NodeJS.Timer, task: () => void, interval: number) => {
-  stopTask(intervalId)
-  return setInterval(task, interval * 1000);
-};
-
-/**
- * Stop the task/interval with the given id.
- *
- * @param intervalId the id of the interval to stop
- * @returns {null} null to reset intervalId
- */
-const stopTask = (intervalId: NodeJS.Timer): null => {
-  if (intervalId) {
-    clearInterval(intervalId);
-  }
-  return null;
-};
-
-/**
- * Normalize values in array (between 0 and 1).
- *
- * @param values the array to normalize
- * @returns {*[]} the array with the normalized values
- */
-const getNormalizedValues = (values: number[]) => {
-  // TODO [#21] always scale y with 0 as lowest, needed/useful if offset/initialSubCount is added?
-  let min = Math.min.apply(Math, values);
-  let max = Math.max.apply(Math, values);
-
-  let normalized = [];
-  for (let value of values) {
-    if (Number.isInteger(value)) {
-      normalized.push((value - min) / (max - min));
-    }
-  }
-  return normalized;
 };
 
 /**
